some refactor and addition of enemy - broken
